+++ AutomatedLab/AutomatedLabAzure.psm1
@@ -87,6 +87,80 @@ function Install-LabAzureRequiredModule
     }
 }
 
+function Register-LabAzureRequiredResourceProvider
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [string]
+        $SubscriptionName,
+
+        [Parameter()]
+        [int]
+        $ProgressIndicator = 5,
+
+        [Parameter()]
+        [switch]
+        $NoDisplay
+    )
+
+    Write-LogFunctionEntry
+
+    $null = Set-AzContext -Subscription $SubscriptionName
+
+    $providers = @(
+        'Microsoft.Network'
+        'Microsoft.Compute'
+        'Microsoft.Storage'
+    )
+
+    $providerObjects = Get-AzResourceProvider -ProviderNamespace $providers | Where-Object RegistrationState -ne 'Registered'
+    if ($providerObjects)
+    {
+        Write-ScreenInfo -Message "Registering required Azure Resource Providers"
+        $providerRegistrations = $providerObjects | Register-AzResourceProvider -ConsentToPermissions $true
+        while ($providerRegistrations.RegistrationState -contains 'Registering')
+        {
+            $providerRegistrations = $providerRegistrations | Get-AzResourceProvider | Where-Object RegistrationState -ne 'Registered'
+            Start-Sleep -Seconds 10
+            Write-ProgressIndicator
+        }
+    }
+
+    $providersAndFeatures = @{
+        'Microsoft.Network' = @(
+            'AllowBastionHost'
+        )
+    }
+
+    $featureState = foreach ($paf in $providersAndFeatures.GetEnumerator())
+    {
+        foreach ($featureName in $paf.Value)
+        {
+            $feature = Get-AzProviderFeature -FeatureName $featureName -ProviderNamespace $paf.Key
+            if ($feature.RegistrationState -eq 'NotRegistered')
+            {
+                Register-AzProviderFeature -FeatureName $featureName -ProviderNamespace $paf.Key
+            }
+        }
+    }
+
+    if (-not $featureState) { Write-LogFunctionExit; return }
+
+    Write-ScreenInfo -Message "Waiting for $($featureState.Count) provider features to register"
+    while ($featureState.RegistrationState -contains 'Registering')
+    {
+        $featureState = $featureState | ForEach-Object {
+            Get-AzProviderFeature -FeatureName $_.FeatureName -ProviderNamespace $_.ProviderName
+        }
+        Start-Sleep -Seconds 10
+        Write-ProgressIndicator
+    }
+
+    Write-LogFunctionExit
+}
+
 function Update-LabAzureSettings
 {
     [CmdletBinding()]
@@ -201,8 +275,8 @@ function Add-LabAzureSubscription
         [void](Set-AzContext -Subscription $SubscriptionId -ErrorAction SilentlyContinue)
     }
 
-    $AzureRmProfile = Get-AzContext
-    if (-not $AzureRmProfile)
+    $azProfile = Get-AzContext
+    if (-not $azProfile)
     {
         throw 'Cannot continue without a valid Azure connection.'
     }
@@ -250,7 +324,7 @@ function Add-LabAzureSubscription
     #select default subscription subscription
     $selectedSubscription = if (-not $SubscriptionName -and -not $SubscriptionId)
     {
-        $AzureRmProfile.Subscription
+        $azProfile.Subscription
     }
     elseif ($SubscriptionName)
     {
@@ -267,6 +341,8 @@ function Add-LabAzureSubscription
     }
 
     Write-ScreenInfo -Message "Using Azure Subscription '$($selectedSubscription.Name)' ($($selectedSubscription.Id))" -Type Info
+
+    Register-LabAzureRequiredResourceProvider -SubscriptionName $selectedSubscription.Name
 
     try
     {
@@ -1543,6 +1619,45 @@ function Get-LabAzureAvailableSku
     
     $publishers |
     Where-Object PublisherName -eq 'MicrosoftWindowsServer' |
+    Get-AzVMImageOffer |
+    Get-AzVMImageSku |
+    Get-AzVMImage |
+    Group-Object -Property Skus, Offer |
+    ForEach-Object { $_.Group | Sort-Object -Property PublishedDate -Descending | Select-Object -First 1 }
+
+    # Linux
+    # Ubuntu - official
+    $publishers |
+    Where-Object PublisherName -eq 'Canonical' |
+    Get-AzVMImageOffer |
+    Where-Object Offer -match '0001-com-ubuntu-server-\w+$' |
+    Get-AzVMImageSku |
+    Where-Object Skus -notmatch 'arm64' |
+    Get-AzVMImage |
+    Group-Object -Property Skus, Offer |
+    ForEach-Object { $_.Group | Sort-Object -Property PublishedDate -Descending | Select-Object -First 1 }
+    # RedHat - official
+    $publishers |
+    Where-Object PublisherName -eq 'RedHat' |
+    Get-AzVMImageOffer |
+    Where-Object Offer -eq 'RHEL' |
+    Get-AzVMImageSku |
+    Where-Object Skus -notmatch '(RAW|LVM|CI)' |
+    Get-AzVMImage |
+    Group-Object -Property Skus, Offer |
+    ForEach-Object { $_.Group | Sort-Object -Property PublishedDate -Descending | Select-Object -First 1 }
+    # CentOS - Roguewave, sounds slightly suspicious
+    $publishers |
+    Where-Object PublisherName -eq 'OpenLogic' |
+    Get-AzVMImageOffer |
+    Where-Object Offer -eq CentOS |
+    Get-AzVMImageSku |
+    Get-AzVMImage |
+    Group-Object -Property Skus, Offer |
+    ForEach-Object { $_.Group | Sort-Object -Property PublishedDate -Descending | Select-Object -First 1 }
+    # Kali
+    $publishers |
+    Where-Object PublisherName -eq 'Kali-Linux' |
     Get-AzVMImageOffer |
     Get-AzVMImageSku |
     Get-AzVMImage |